refactor: extract instance from MainWindow into InstanceService
Move backup, clone, create, delete, and restore instance operations
out of MainWindow into a dedicated self-contained service layer.

Key changes:
- New app/services/instance_service.py: self-subscribes to EventBus,
  handles all instance lifecycle orchestration with minimal constructor
  deps (settings, steamcmd_wrapper)
- main_window.py: -786 LOC — removed 5 private methods with nested
  helpers and 3 dead dialog methods; 5 unused imports cleaned
- app_controller.py: creates InstanceService in its own init step
- InstanceService uses EventBus signals and QInputDialog directly
  instead of MainWindow callbacks, making it fully independent
