+++ app/views/main_window.py
@@ -1,10 +1,7 @@
 import os
 from collections.abc import Callable
 from functools import partial
-from pathlib import Path
-from shutil import copytree, rmtree
-from traceback import format_exc
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
     from app.views.settings_dialog import SettingsDialog
@@ -15,10 +12,8 @@ from PySide6.QtGui import QCloseEvent, QShowEvent
 from PySide6.QtWidgets import (
     QApplication,
     QHBoxLayout,
-    QInputDialog,
     QLabel,
     QMainWindow,
-    QMessageBox,
     QPushButton,
     QTabWidget,
     QVBoxLayout,
@@ -26,14 +21,10 @@ from PySide6.QtWidgets import (
 )
 
 from app.controllers.file_search_controller import FileSearchController
-from app.controllers.instance_controller import (
-    InstanceController,
-    InvalidArchivePathError,
-)
+from app.controllers.instance_controller import InstanceController
 from app.controllers.main_content_controller import MainContentController
 from app.controllers.menu_bar_controller import MenuBarController
 from app.controllers.metadata_controller import MetadataController
-from app.controllers.metadata_db_controller import AuxMetadataController
 from app.controllers.mods_panel_controller import ModsPanelController
 from app.controllers.todds_controller import ToddsController
 from app.controllers.troubleshooting_controller import TroubleshootingController
@@ -42,26 +33,13 @@ from app.models.settings import Settings
 from app.utils import globals
 from app.utils.acf_utils import refresh_acf_metadata
 from app.utils.app_info import AppInfo
-from app.utils.constants import (
-    DEFAULT_INSTANCE_NAME,
-    INSTANCE_FOLDER_NAME,
-    STEAM_FOLDER_NAME,
-    STEAMCMD_FOLDER_NAME,
-)
 from app.utils.event_bus import EventBus
-from app.utils.generic import handle_remove_read_only
 from app.utils.gui_info import GUIInfo
 from app.utils.steam.steamcmd.wrapper import SteamcmdInterface
 from app.utils.watchdog import WatchdogHandler
 from app.utils.window_launch_state import apply_window_launch_state
 from app.views.acf_log_reader import AcfLogReader
-from app.views.dialogue import (
-    BinaryChoiceDialog,
-    show_dialogue_conditional,
-    show_dialogue_file,
-    show_fatal_error,
-    show_warning,
-)
+from app.views.dialogue import BinaryChoiceDialog
 from app.views.file_search_dialog import FileSearchDialog
 from app.views.main_content_panel import MainContent
 from app.views.menu_bar import MenuBar
@@ -133,6 +111,7 @@ class MainWindow(QMainWindow):
             self.__disable_enable_widgets
         )
         self.main_content_panel.stop_watchdog_signal.connect(self.shutdown_watchdog)
+
         self.bottom_panel = Status()
 
         # Create and add the Main Content panel tab
@@ -265,13 +244,6 @@ class MainWindow(QMainWindow):
 
         # Connect Instances Menu Bar signals
         EventBus().do_activate_current_instance.connect(self.__switch_to_instance)
-        EventBus().do_backup_existing_instance.connect(self.__backup_existing_instance)
-        EventBus().do_clone_existing_instance.connect(self.__clone_existing_instance)
-        EventBus().do_create_new_instance.connect(self.__create_new_instance)
-        EventBus().do_delete_current_instance.connect(self.__delete_current_instance)
-        EventBus().do_restore_instance_from_archive.connect(
-            self.__restore_instance_from_archive
-        )
 
         # launch the main window
         self._launch_main_window()
@@ -394,756 +366,6 @@ class MainWindow(QMainWindow):
                 self._set_instance(instance)
 
         return
-
-    def __ask_for_new_instance_name(self) -> str | None:
-        instance_name, ok = QInputDialog.getText(
-            self,
-            self.tr("Create new instance"),
-            self.tr("Input a unique name of new instance that is not already used:"),
-        )
-        return instance_name.strip() if ok else None
-
-    def __ask_for_non_default_instance_name(self) -> str | None:
-        while True:
-            instance_name, ok = QInputDialog.getText(
-                self,
-                self.tr("Provide instance name"),
-                self.tr(
-                    'Input a unique name for the backed up instance that is not "{name}"'
-                ).format(name=DEFAULT_INSTANCE_NAME),
-            )
-            if ok and instance_name.lower() != DEFAULT_INSTANCE_NAME.lower():
-                return instance_name
-            else:
-                return None
-
-    def __ask_how_to_workshop_mods(
-        self, existing_instance_name: str, existing_instance_workshop_folder: str
-    ) -> str:
-        answer = show_dialogue_conditional(
-            title=self.tr("Clone instance [{existing_instance_name}]").format(
-                existing_instance_name=existing_instance_name
-            ),
-            text=(
-                self.tr(
-                    "What would you like to do with the configured Workshop mods folder?"
-                )
-            ),
-            information=(
-                self.tr(
-                    "Workshop folder: {existing_instance_workshop_folder}<br><br>"
-                    + "Option 1: Convert to SteamCMD<br>"
-                    + "RimSort will copy all Workshop mods to the new instance's local mods folder, converting them to SteamCMD mods that you can manage inside the new instance. The Workshop folder will be ignored for this instance to prevent duplicate mods.<br><br>"
-                    + "Option 2: Keep Workshop Folder<br>"
-                    + "The new instance will use the same Workshop folder as the original instance. You can change this later in the settings if needed.<br><br>"
-                    + "How would you like to proceed?"
-                ).format(
-                    existing_instance_workshop_folder=existing_instance_workshop_folder
-                )
-            ),
-            button_text_override=[
-                self.tr("Convert to SteamCMD"),
-                self.tr("Keep Workshop Folder"),
-            ],
-        )
-        answer_str = str(answer)
-        cancelled_str = self.tr("Cancelled")
-        return answer_str or cancelled_str
-
-    def __backup_existing_instance(self, instance_name: str) -> None:
-        """Backup an instance to a ZIP archive."""
-        # Get instance data from Settings
-        instance = self.settings.instances.get(instance_name)
-
-        # If the instance_name is "Default", prompt the user for a new instance name.
-        if instance_name == DEFAULT_INSTANCE_NAME:
-            new_instance_name = self.__ask_for_non_default_instance_name()
-            if not new_instance_name:
-                logger.info("User cancelled operation")
-                return
-            instance_name = new_instance_name
-
-        # Determine instance data to save
-        if instance is None:
-            logger.error(f"Instance [{instance_name}] not found in Settings")
-            return
-
-        instance_controller = InstanceController(instance)
-        # Prompt user to select output path for instance archive
-        output_path = show_dialogue_file(
-            mode="save",
-            caption="Select output path for instance archive",
-            _dir=str(AppInfo().app_storage_folder),
-            _filter="Zip files (*.zip)",
-        )
-        logger.info(f"Selected path: {output_path}")
-        if output_path:
-            try:
-                EventBus().do_threaded_loading_animation.emit(
-                    str(AppInfo().theme_data_folder / "default-icons" / "rimsort.gif"),
-                    partial(
-                        instance_controller.compress_to_archive,
-                        output_path,
-                    ),
-                    self.tr(
-                        "Compressing [{instance_name}] instance folder to archive..."
-                    ).format(instance_name=instance_name),
-                )
-            except Exception as e:
-                show_fatal_error(
-                    title=self.tr("Error compressing instance"),
-                    text=self.tr(
-                        "An error occurred while compressing instance folder: {e}"
-                    ).format(e=e),
-                    information=self.tr("Please check the logs for more information."),
-                    details=format_exc(),
-                )
-        else:
-            logger.warning("Backup cancelled: User cancelled selection...")
-            return
-
-    def __restore_instance_from_archive(self) -> None:
-        """Restore an instance from a ZIP archive."""
-        # Prompt user to select input path for instance archive
-        input_path = show_dialogue_file(
-            mode="open",
-            caption="Select input path for instance archive",
-            _dir=str(AppInfo().app_storage_folder),
-            _filter="Zip files (*.zip)",
-        )
-
-        if input_path is None:
-            logger.info("User cancelled operation. Input path was None")
-            return
-        logger.info(f"Selected path: {input_path}")
-
-        if not os.path.exists(input_path):
-            logger.error(f"Archive not found at path: {input_path}")
-            show_warning(
-                title=self.tr("Error restoring instance"),
-                text=self.tr("Archive not found at path: {input_path}").format(
-                    input_path=input_path
-                ),
-            )
-            return
-
-        # Load instance from archive using factory method
-        try:
-            instance_controller = InstanceController.from_archive(input_path)
-        except InvalidArchivePathError:
-            # Handled in controller. Gracefully fail.
-            return
-        except Exception as e:
-            logger.error(f"An error occurred while reading instance archive: {e}")
-            show_fatal_error(
-                title=self.tr("Error restoring instance"),
-                text=self.tr("An error occurred while reading instance archive: {e}"),
-                details=format_exc(),
-            )
-            return
-
-        if os.path.exists(instance_controller.instance_folder_path):
-            answer = show_dialogue_conditional(
-                title=self.tr("Instance folder exists"),
-                text=self.tr(
-                    "Instance folder already exists: {instance_folder_path}"
-                ).format(instance_folder_path=instance_controller.instance_folder_path),
-                information=self.tr(
-                    "Do you want to continue and replace the existing instance folder?"
-                ),
-                button_text_override=[
-                    self.tr("Replace"),
-                ],
-            )
-
-            if answer != self.tr("Replace"):
-                logger.info("User cancelled instance extraction.")
-                return
-
-        EventBus().do_threaded_loading_animation.emit(
-            str(AppInfo().theme_data_folder / "default-icons" / "rimsort.gif"),
-            partial(
-                instance_controller.extract_from_archive,
-                input_path,
-            ),
-            self.tr("Restoring instance [{name}] from archive...").format(
-                name=instance_controller.instance.name
-            ),
-        )
-
-        # Check that the instance folder exists. If it does, update Settings with the instance data
-        if os.path.exists(instance_controller.instance_folder_path):
-            cleared_paths = instance_controller.validate_paths()
-            if cleared_paths:
-                logger.warning(
-                    f"Instance folder paths not found: {', '.join(cleared_paths)}"
-                )
-                show_warning(
-                    title=self.tr("Invalid instance folder paths"),
-                    text=self.tr("Invalid instance folder paths"),
-                    information=self.tr(
-                        "Some folder paths from the restored instance are invalid and were cleared. Please reconfigure them in the settings"
-                    ),
-                    details=self.tr("Invalid paths: {path}").format(
-                        path=", ".join(cleared_paths)
-                    ),
-                )
-
-            steamcmd_link_path = str(
-                Path(instance_controller.instance.steamcmd_install_path)
-                / "steam"
-                / "steamapps"
-                / "workshop"
-                / "content"
-                / "294100"
-            )
-
-            if (
-                os.path.exists(steamcmd_link_path)
-                and instance_controller.instance.local_folder != ""
-            ):
-                logger.info("Restoring steamcmd symlink...")
-                self.steamcmd_wrapper.create_symlink(
-                    instance_controller.instance.local_folder,
-                    steamcmd_link_path,
-                    show_dialogues=False,
-                    force=True,
-                )
-            elif not os.path.exists(steamcmd_link_path):
-                logger.info("Skipping steamcmd symlink restoration")
-            else:
-                show_warning(
-                    title=self.tr("Couldn't restore steamcmd symlink/junction"),
-                    text=self.tr("Couldn't restore steamcmd symlink/junction"),
-                    information=self.tr(
-                        "The steamcmd symlink/junction could not be restored as the local folder is not set or invalid. The symlink/junction will need to be manually recreated."
-                    ),
-                )
-                logger.warning(
-                    "Skipping steamcmd symlink restoration: Local folder not set. The symlink will need to be manually updated."
-                )
-
-            self._set_instance(instance_controller.instance)
-            self.__switch_to_instance(instance_controller.instance.name)
-        else:
-            show_warning(
-                title=self.tr("Error restoring instance"),
-                text=self.tr(
-                    "An error occurred while restoring instance [{name}]."
-                ).format(name=instance_controller.instance.name),
-                information=self.tr(
-                    "The instance folder was not found after extracting the archive. Perhaps the archive is corrupt or the instance name is invalid."
-                ),
-            )
-
-            logger.warning(
-                "Restore cancelled: Instance folder not found after extraction..."
-            )
-
-    def __clone_existing_instance(self, existing_instance_name: str) -> None:
-        """Clone an existing instance to create a new one with copied data."""
-
-        def copy_game_folder(
-            existing_instance_game_folder: str, target_game_folder: str
-        ) -> None:
-            try:
-                if os.path.exists(target_game_folder) and os.path.isdir(
-                    target_game_folder
-                ):
-                    logger.info(
-                        f"Replacing existing game folder at {target_game_folder}"
-                    )
-                    rmtree(
-                        target_game_folder,
-                        ignore_errors=False,
-                        onerror=handle_remove_read_only,
-                    )
-                logger.info(
-                    f"Copying game folder from {existing_instance_game_folder} to {target_game_folder}"
-                )
-                copytree(
-                    existing_instance_game_folder, target_game_folder, symlinks=True
-                )
-            except Exception as e:
-                logger.error(f"An error occurred while copying game folder: {e}")
-
-        def copy_config_folder(
-            existing_instance_config_folder: str, target_config_folder: str
-        ) -> None:
-            try:
-                if os.path.exists(target_config_folder) and os.path.isdir(
-                    target_config_folder
-                ):
-                    logger.info(
-                        f"Replacing existing config folder at {target_config_folder}"
-                    )
-                    rmtree(
-                        target_config_folder,
-                        ignore_errors=False,
-                        onerror=handle_remove_read_only,
-                    )
-                logger.info(
-                    f"Copying config folder from {existing_instance_config_folder} to {target_config_folder}"
-                )
-                copytree(
-                    existing_instance_config_folder,
-                    target_config_folder,
-                    symlinks=True,
-                )
-            except Exception as e:
-                logger.error(f"An error occurred while copying config folder: {e}")
-
-        def copy_local_folder(
-            existing_instance_local_folder: str, target_local_folder: str
-        ) -> None:
-            try:
-                if os.path.exists(target_local_folder) and os.path.isdir(
-                    target_local_folder
-                ):
-                    logger.info(
-                        f"Replacing existing local folder at {target_local_folder}"
-                    )
-                    rmtree(
-                        target_local_folder,
-                        ignore_errors=False,
-                        onerror=handle_remove_read_only,
-                    )
-                logger.info(
-                    f"Copying local folder from {existing_instance_local_folder} to {target_local_folder}"
-                )
-                copytree(
-                    existing_instance_local_folder,
-                    target_local_folder,
-                    symlinks=True,
-                )
-            except Exception as e:
-                logger.error(f"An error occurred while copying local folder: {e}")
-
-        def copy_workshop_mods_to_local(
-            existing_instance_workshop_folder: str, target_local_folder: str
-        ) -> None:
-            try:
-                if not os.path.exists(target_local_folder):
-                    os.mkdir(target_local_folder)
-                logger.info(
-                    f"Cloning Workshop mods from {existing_instance_workshop_folder} to {target_local_folder}"
-                )
-                # Copy each subdirectory of the existing Workshop folder to the new local mods folder
-                for subdir in os.listdir(existing_instance_workshop_folder):
-                    if os.path.isdir(
-                        os.path.join(existing_instance_workshop_folder, subdir)
-                    ):
-                        logger.debug(f"Cloning Workshop mod: {subdir}")
-                        copytree(
-                            os.path.join(existing_instance_workshop_folder, subdir),
-                            os.path.join(target_local_folder, subdir),
-                            symlinks=True,
-                        )
-            except Exception as e:
-                logger.error(f"An error occurred while cloning Workshop mods: {e}")
-
-        def clone_essential_paths(
-            existing_instance_game_folder: str,
-            target_game_folder: str,
-            existing_instance_config_folder: str,
-            target_config_folder: str,
-        ) -> None:
-            # Clone the existing game_folder to the new instance
-            if os.path.exists(existing_instance_game_folder) and os.path.isdir(
-                existing_instance_game_folder
-            ):
-                copy_game_folder(existing_instance_game_folder, target_game_folder)
-            # Clone the existing config_folder to the new instance
-            if os.path.exists(existing_instance_config_folder) and os.path.isdir(
-                existing_instance_config_folder
-            ):
-                copy_config_folder(
-                    existing_instance_config_folder, target_config_folder
-                )
-
-        # Check if paths are set. We can't clone if they aren't set
-        if not self.main_content_panel.check_if_essential_paths_are_set(prompt=True):
-            return
-        # Get instance data from Settings
-        current_instances = list(self.settings.instances.keys())
-        existing_instance = self.settings.instances[existing_instance_name]
-
-        existing_instance_game_folder = existing_instance.game_folder
-        game_folder_name = os.path.split(existing_instance_game_folder)[1]
-        existing_instance_local_folder = existing_instance.local_folder
-        local_folder_name = os.path.split(existing_instance_local_folder)[1]
-        existing_instance_workshop_folder = existing_instance.workshop_folder
-        existing_instance_config_folder = existing_instance.config_folder
-        existing_instance_run_args = existing_instance.run_args
-        existing_instance_steamcmd_install_path = (
-            existing_instance.steamcmd_install_path
-        )
-        existing_instance_steam_client_integration = (
-            existing_instance.steam_client_integration
-        )
-        existing_instance_folder_override = existing_instance.instance_folder_override
-        # Sanitize the input so that it does not produce any KeyError down the road
-        new_instance_name = self.__ask_for_new_instance_name()
-        if (
-            new_instance_name
-            and new_instance_name != DEFAULT_INSTANCE_NAME
-            and new_instance_name not in current_instances
-        ):
-            new_instance_path = str(
-                InstanceController.get_instance_folder_path(
-                    new_instance_name, existing_instance_folder_override
-                )
-            )
-            # Prompt user with the existing instance configuration and confirm that they would like to clone it
-            answer = BinaryChoiceDialog(
-                title=f"Clone instance [{existing_instance_name}]",
-                text=f"Would you like to clone instance [{existing_instance_name}] to create new instance [{new_instance_name}]?<br>"
-                + "<br>This will clone the instance's game, mod, and configuration data. This operation may take a long time depending on the amount of data being cloned.<br>"
-                + "<br>The following folders will be cloned:",
-                information=f"Game folder:<br>{existing_instance_game_folder if existing_instance_game_folder else '&lt;None&gt;'}<br>"
-                + f"<br>Configuration folder:<br>{existing_instance_config_folder if existing_instance_config_folder else '&lt;None&gt;'}<br>"
-                + f"<br>Local mods folder:<br>{existing_instance_local_folder if existing_instance_local_folder else '&lt;None&gt;'}<br>"
-                + f"<br>Workshop mods folder:<br>{existing_instance_workshop_folder if existing_instance_workshop_folder else '&lt;None&gt;'}<br>"
-                + "<br>SteamCMD install path (steamcmd + steam folders will be cloned):"
-                + f"<br>{existing_instance_steamcmd_install_path if existing_instance_steamcmd_install_path else '&lt;None&gt;'}<br>"
-                + f"<br>Run arguments:<br>{existing_instance_run_args if existing_instance_run_args else '&lt;None&gt;'}<br>",
-            )
-            if answer.exec_is_positive():
-                # Clone the RimWorld game_folder to the new instance
-                target_game_folder = str(Path(new_instance_path) / game_folder_name)
-                target_local_folder = str(
-                    Path(new_instance_path) / game_folder_name / local_folder_name
-                )
-                target_workshop_folder = ""
-                target_config_folder = str(
-                    Path(new_instance_path) / "InstanceData" / "Config"
-                )
-                EventBus().do_threaded_loading_animation.emit(
-                    str(AppInfo().theme_data_folder / "default-icons" / "rimworld.gif"),
-                    partial(
-                        clone_essential_paths,
-                        existing_instance_game_folder,
-                        target_game_folder,
-                        existing_instance_config_folder,
-                        target_config_folder,
-                    ),
-                    f"Cloning RimWorld game / config folders from [{existing_instance_name}] to [{new_instance_name}] instance...",
-                )
-                # Clone the existing local_folder to the new instance
-                # Skip if local folder is already included within the game folder (prevents redundant cloning)
-                local_folder_in_game = (
-                    str(Path(existing_instance_game_folder) / local_folder_name)
-                    == existing_instance_local_folder
-                )
-                if existing_instance_local_folder and not local_folder_in_game:
-                    if os.path.exists(existing_instance_local_folder) and os.path.isdir(
-                        existing_instance_local_folder
-                    ):
-                        EventBus().do_threaded_loading_animation.emit(
-                            str(
-                                AppInfo().theme_data_folder
-                                / "default-icons"
-                                / "rimworld.gif"
-                            ),
-                            partial(
-                                copy_local_folder,
-                                existing_instance_local_folder,
-                                target_local_folder,
-                            ),
-                            f"Cloning local mods folder from [{existing_instance_name}] instance to [{new_instance_name}] instance...",
-                        )
-                # Clone the existing workshop_folder to the new instance's local mods folder
-                if existing_instance_workshop_folder:
-                    # Prompt user to confirm before initiating the procedure
-                    answer_workshop_mods: str = self.__ask_how_to_workshop_mods(
-                        existing_instance_name=existing_instance_name,
-                        existing_instance_workshop_folder=existing_instance_workshop_folder,
-                    )
-                    if answer_workshop_mods == self.tr("Convert to SteamCMD"):
-                        if os.path.exists(
-                            existing_instance_workshop_folder
-                        ) and os.path.isdir(existing_instance_workshop_folder):
-                            EventBus().do_threaded_loading_animation.emit(
-                                str(
-                                    AppInfo().theme_data_folder
-                                    / "default-icons"
-                                    / "steam_api.gif"
-                                ),
-                                partial(
-                                    copy_workshop_mods_to_local,
-                                    existing_instance_workshop_folder,
-                                    target_local_folder,
-                                ),
-                                f"Cloning Workshop mods from [{existing_instance_name}] instance to [{new_instance_name}] instance's local mods...",
-                            )
-                        else:
-                            show_warning(
-                                title=self.tr("Workshop mods not found"),
-                                text=self.tr(
-                                    "Workshop mods folder at [{existing_instance_workshop_folder}] not found."
-                                ).format(
-                                    existing_instance_workshop_folder=existing_instance_workshop_folder
-                                ),
-                            )
-                    elif answer_workshop_mods == self.tr("Keep Workshop Folder"):
-                        target_workshop_folder = str(existing_instance_workshop_folder)
-                # If the instance has a 'steamcmd' folder, clone it to the new instance
-                steamcmd_install_path = str(
-                    Path(existing_instance_steamcmd_install_path) / STEAMCMD_FOLDER_NAME
-                )
-                if os.path.exists(steamcmd_install_path) and os.path.isdir(
-                    steamcmd_install_path
-                ):
-                    target_steamcmd_install_path = str(
-                        Path(new_instance_path) / STEAMCMD_FOLDER_NAME
-                    )
-                    if os.path.exists(target_steamcmd_install_path) and os.path.isdir(
-                        target_steamcmd_install_path
-                    ):
-                        logger.info(
-                            f"Replacing existing steamcmd folder at {target_steamcmd_install_path}"
-                        )
-                        rmtree(
-                            target_steamcmd_install_path,
-                            ignore_errors=False,
-                            onerror=handle_remove_read_only,
-                        )
-                    logger.info(
-                        f"Copying steamcmd folder from {steamcmd_install_path} to {target_steamcmd_install_path}"
-                    )
-                    copytree(
-                        steamcmd_install_path,
-                        target_steamcmd_install_path,
-                        symlinks=True,
-                    )
-                # If the instance has a 'steam' folder, clone it to the new instance
-                steam_install_path = str(
-                    Path(existing_instance_steamcmd_install_path) / STEAM_FOLDER_NAME
-                )
-                if os.path.exists(steam_install_path) and os.path.isdir(
-                    steam_install_path
-                ):
-                    target_steam_install_path = str(
-                        Path(new_instance_path) / STEAM_FOLDER_NAME
-                    )
-                    if os.path.exists(target_steam_install_path) and os.path.isdir(
-                        target_steam_install_path
-                    ):
-                        logger.info(
-                            f"Replacing existing steam folder at {target_steam_install_path}"
-                        )
-                        rmtree(
-                            target_steam_install_path,
-                            ignore_errors=False,
-                            onerror=handle_remove_read_only,
-                        )
-                    logger.info(
-                        f"Copying steam folder from {steam_install_path} to {target_steam_install_path}"
-                    )
-                    # Copy the directory, but omit the symlink path
-                    copytree(
-                        steam_install_path,
-                        target_steam_install_path,
-                        symlinks=True,
-                        ignore=lambda d, names: (
-                            ["steamapps/workshop/content/294100"]
-                            if d == steam_install_path
-                            else []
-                        ),
-                    )
-                    # Unlink steam/workshop/content/294100 symlink if it exists, and relink it to our new target local mods folder
-                    link_path = str(
-                        Path(target_steam_install_path)
-                        / "steamapps"
-                        / "workshop"
-                        / "content"
-                        / "294100"
-                    )
-                    self.steamcmd_wrapper.create_symlink(
-                        target_local_folder, link_path, show_dialogues=False, force=True
-                    )
-                # Create the new instance for our cloned instance
-                self.__create_new_instance(
-                    instance_name=new_instance_name,
-                    instance_data={
-                        "game_folder": target_game_folder,
-                        "local_folder": target_local_folder,
-                        "workshop_folder": target_workshop_folder,
-                        "config_folder": target_config_folder,
-                        "run_args": existing_instance_run_args,
-                        "steamcmd_install_path": str(
-                            AppInfo().app_storage_folder
-                            / INSTANCE_FOLDER_NAME
-                            / new_instance_name
-                        ),
-                        "steam_client_integration": existing_instance_steam_client_integration,
-                        "instance_folder_override": existing_instance_folder_override,
-                    },
-                )
-        elif new_instance_name:
-            show_warning(
-                title=self.tr("Error cloning instance"),
-                text=self.tr("Unable to clone instance."),
-                information=self.tr(
-                    f"Please enter a valid, unique instance name. It cannot be '{DEFAULT_INSTANCE_NAME}' or empty."
-                ),
-            )
-        else:
-            logger.debug("User cancelled clone operation")
-
-    def __create_new_instance(
-        self, instance_name: str = "", instance_data: dict[str, Any] | None = None
-    ) -> None:
-        """Create a new instance with the provided name and data."""
-        if instance_data is None:
-            instance_data = {}
-        if not instance_name:
-            # Sanitize the input so that it does not produce any KeyError down the road
-            new_instance_name = self.__ask_for_new_instance_name()
-            if not new_instance_name:
-                logger.info("User cancelled operation")
-                return
-            instance_name = new_instance_name
-        current_instances = list(self.settings.instances.keys())
-        if (
-            instance_name
-            and instance_name != DEFAULT_INSTANCE_NAME
-            and instance_name not in current_instances
-        ):
-            # Get instance folder override if provided
-            # First check if override was passed in instance_data (e.g., from cloning)
-            instance_folder_override = instance_data.get("instance_folder_override", "")
-            # If not provided in instance_data, check the current instance's override
-            # This allows new instances to use the custom folder set in settings
-            if not instance_folder_override:
-                current_instance = self.settings.instances[
-                    self.settings.current_instance
-                ]
-                instance_folder_override = current_instance.instance_folder_override
-            # Create new instance folder if it does not exist
-            instance_path = InstanceController.get_instance_folder_path(
-                instance_name, instance_folder_override
-            )
-            if not instance_path.exists():
-                instance_path.mkdir(parents=True, exist_ok=True)
-            # Fall back to current instance's folders if missing or empty (e.g., Create Instance)
-            current_inst = self.settings.instances[self.settings.current_instance]
-            if not instance_data.get("game_folder"):
-                instance_data["game_folder"] = current_inst.game_folder
-            if not instance_data.get("config_folder"):
-                instance_data["config_folder"] = current_inst.config_folder
-            # Get run args from instance data, autogenerate additional config items if desired
-            run_args = ""
-            if instance_data.get("game_folder") and instance_data.get("config_folder"):
-                # Generate preview of run args as a single string
-                log_path = str(instance_path / "RimWorld.log")
-                savedata_path = str(instance_path / "InstanceData")
-                preview_text = f"-logfile {log_path} -savedatafolder={savedata_path}"
-                # Prompt the user if they would like to automatically generate run args for the instance
-                answer = show_dialogue_conditional(
-                    title=self.tr("Create new instance [{instance_name}]"),
-                    text=self.tr(
-                        "Would you like to automatically generate run args for the new instance?"
-                    ),
-                    information=self.tr(
-                        "This will try to generate run args for the new instance based on the configured Game/Config folders.<br><br>"
-                        + "Generated run arguments preview:<br>{preview}"
-                    ).format(preview=preview_text),
-                )
-                if answer == QMessageBox.StandardButton.Yes:
-                    run_args = preview_text
-                    # Align config_folder with -savedatafolder so _do_save()
-                    # writes ModsConfig.xml where the game will read it
-                    config_path = instance_path / "InstanceData" / "Config"
-                    config_path.mkdir(parents=True, exist_ok=True)
-                    instance_data["config_folder"] = str(config_path)
-                # Append any existing run_args from cloned instance
-                existing_args = instance_data.get("run_args", "")
-                if existing_args:
-                    run_args = f"{run_args} {existing_args}".strip()
-            # Add new instance to Settings
-            instance = InstanceController.create_instance(
-                instance_name=instance_name,
-                game_folder=instance_data.get("game_folder", ""),
-                local_folder=instance_data.get("local_folder", ""),
-                workshop_folder=instance_data.get("workshop_folder", ""),
-                config_folder=instance_data.get("config_folder", ""),
-                run_args=run_args,
-                steamcmd_install_path=str(instance_path),
-                steam_client_integration=instance_data.get(
-                    "steam_client_integration", False
-                ),
-                instance_folder_override=instance_folder_override,
-            )
-            self._set_instance(instance)
-
-            # Save settings
-            self.settings.save()
-            # Switch to new instance and initialize content
-            self.__switch_to_instance(instance_name)
-        else:
-            show_warning(
-                title=self.tr("Error creating instance"),
-                text=self.tr("Unable to create new instance."),
-                information=self.tr(
-                    f"Please enter a valid, unique instance name. It cannot be '{DEFAULT_INSTANCE_NAME}' or empty."
-                ),
-            )
-
-    def __delete_current_instance(self) -> None:
-        """Delete the current instance and all its data."""
-        if self.settings.current_instance == DEFAULT_INSTANCE_NAME:
-            show_warning(
-                title=self.tr("Problem deleting instance"),
-                text=self.tr("Unable to delete instance {current_instance}.").format(
-                    current_instance=self.settings.current_instance
-                ),
-                information=self.tr("The default instance cannot be deleted."),
-            )
-            return
-        elif not self.settings.instances.get(self.settings.current_instance):
-            show_fatal_error(
-                title=self.tr("Error deleting instance"),
-                text=self.tr("Unable to delete instance {current_instance}.").format(
-                    current_instance=self.settings.current_instance
-                ),
-                information=self.tr("The selected instance does not exist."),
-            )
-            return
-        else:
-            answer = BinaryChoiceDialog(
-                title=self.tr("Delete instance {current_instance}").format(
-                    current_instance=self.settings.current_instance
-                ),
-                text=self.tr(
-                    "Are you sure you want to delete the selected instance and all of its data?"
-                ),
-                information=self.tr("This action cannot be undone."),
-            )
-            if answer.exec_is_positive():
-                aux_metadata_controller = (
-                    AuxMetadataController.get_or_create_cached_instance(
-                        self.settings.aux_db_path
-                    )
-                )
-                aux_metadata_controller.engine.dispose()
-                try:
-                    rmtree(
-                        str(
-                            Path(
-                                AppInfo().app_storage_folder
-                                / INSTANCE_FOLDER_NAME
-                                / self.settings.current_instance
-                            )
-                        ),
-                        ignore_errors=False,
-                        onerror=handle_remove_read_only,
-                    )
-                except Exception as e:
-                    logger.error(f"Error deleting instance: {e}")
-                # Remove instance from settings and reset to Default
-                self.settings.instances.pop(self.settings.current_instance)
-                self.__switch_to_instance(DEFAULT_INSTANCE_NAME)
 
     def __switch_to_instance(self, instance: str) -> None:
         """Switch to a different instance."""
